Refactor imports and types
