--- src/providers/Theme/ThemeSelector/types.ts
@@ -1,5 +1,0 @@
-export type Theme = "dark" | "light";
-
-export const themeLocalStorageKey = "site-theme";
-
-export const defaultTheme = "light";
